IProduct and it's implementation are added

+++ src/main/java/tn/talan/internship/dto/ProductDTO.java
@@ -2,14 +2,16 @@ package tn.talan.internship.dto;
 
 import lombok.Data;
 import org.modelmapper.ModelMapper;
+import org.springframework.beans.BeanUtils;
 
 import java.io.Serializable;
 
 @Data
 public class ProductDTO implements Serializable {
+
     private Long id ;
     private String name;
     private Double sellPrice;
-    private int quantity;
+    private Long quantity;
 
 }
